requirements.txt
facebook-sdk=3.1.0
pandas=1.4.3
requests=2.28.1